Add files via upload
It now takes the csv, makes a hashmap from it, and writes the hashmap to a file. Then when wordle runs it iterates through that map and puts all the words of n length in a new map and writes that to a file. However, if that hashmap already exists it will just use that. 
I also put the script in a loop that runs until the score is all 2s so you can actually play wordle. 
The getInput() function could be overwritten by an algorithm that picks inputs.